requirements.txt
beautifulsoup4
discord.py==1.7.3
python-dotenv